+++ packages/meta/src/utils/generation/internal.ts
@@ -7,4 +7,4 @@ export * from "./LangUtil.js";
 export * from "./Roles.js";
 export * from "./FileWatcher.js";
 export * from "./ReservedWords.js";
-export * from "./ImportsUtil.js"
+export * from "./Imports.js"
